chore: moved and renamed Omnichannel's voip files
