--- apps/meteor/client/providers/CallProvider/index.ts
@@ -1,1 +1,0 @@
-export * from './CallProvider';
