+++ apps/meteor/client/providers/OmnichannelCallProvider/index.ts
@@ -1,0 +1,1 @@
+export { default } from './OmnichannelCallProvider';
