+++ apps/meteor/client/sidebar/footer/voip/OmnichannelVoipFooter.tsx
@@ -6,8 +6,8 @@ import { useTranslation } from '@rocket.chat/ui-contexts';
 import type { ReactElement, MouseEvent, ReactNode } from 'react';
 import React from 'react';
 
-import type { VoipFooterMenuOptions } from '../../../../ee/client/hooks/useVoipFooterMenu';
-import type { CallActionsType } from '../../../contexts/CallContext';
+import type { VoipFooterMenuOptions } from '../../../../ee/client/hooks/useOmnichannelVoipFooterMenu';
+import type { CallActionsType } from '../../../contexts/OmnichannelCallContext';
 import { useOmnichannelContactLabel } from './hooks/useOmnichannelContactLabel';
 
 type VoipFooterPropsType = {
