+++ apps/meteor/client/views/omnichannel/directory/calls/contextualBar/VoipInfoCallButton.tsx
@@ -3,7 +3,7 @@ import { useTranslation } from '@rocket.chat/ui-contexts';
 import type { ComponentProps, ReactElement } from 'react';
 import React from 'react';
 
-import { useVoipOutboundStates } from '../../../../../contexts/CallContext';
+import { useVoipOutboundStates } from '../../../../../contexts/OmnichannelCallContext';
 import { useDialModal } from '../../../../../hooks/useDialModal';
 
 type VoipInfoCallButtonProps = Exclude<
