+++ apps/meteor/ee/client/hooks/useOmnichannelOutboundDialer.ts
@@ -1,7 +1,7 @@
-import { useCallClient, useIsVoipEnterprise } from '../../../client/contexts/CallContext';
+import { useCallClient, useIsVoipEnterprise } from '../../../client/contexts/OmnichannelCallContext';
 import { EEVoipClient } from '../lib/voip/EEVoipClient';
 
-export const useOutboundDialer = (): EEVoipClient | null => {
+export const useOmnichannelOutboundDialer = (): EEVoipClient | null => {
 	const voipClient = useCallClient();
 	const isEnterprise = useIsVoipEnterprise();
 	const isOutboundClient = voipClient instanceof EEVoipClient;
