+++ apps/meteor/tests/unit/client/providers/CallProvider/lib/parseStringToIceServers.spec.ts
@@ -1,7 +1,7 @@
 import {
 	parseStringToIceServers,
 	parseStringToIceServer,
-} from '../../../../../../client/providers/CallProvider/lib/parseStringToIceServers';
+} from '../../../../../../client/providers/OmnichannelCallProvider/lib/parseStringToIceServers';
 
 describe('parseStringToIceServers', () => {
 	describe('parseStringToIceServers', () => {
